Fix text overlay export crash: accept float values for outline_width, font_size, font_weight
The UI allows fractional values (e.g. outline_width=5.5) but the backend
Pydantic model required strict int types, causing a 422 validation error
on export. Changed TextOverlay model fields to float and added rounding
in both the FFmpeg filter builder (backend) and the payload builder
(frontend) to ensure valid integer values reach FFmpeg commands.

+++ backend/models.py
@@ -173,12 +173,12 @@ class TextOverlay(BaseModel):
     text: str = ""
     x: float = 50              # position % (0-100)
     y: float = 50              # position % (0-100)
-    font_size: int = 48
+    font_size: float = 48
     font_color: str = "#FFFFFF"
     font_family: str = "sans-serif"
-    font_weight: int = 400
+    font_weight: float = 400
     background_color: Optional[str] = None
-    outline_width: int = 0
+    outline_width: float = 0
     outline_color: str = "#000000"
     start_time: float = 0.0    # relative to clip start
     end_time: float = 0.0
